refactor: update component structure
CU-3cy5cnn
