+++ src/components/controls/Checkbox.js
@@ -5,10 +5,29 @@ import {
 } from "@mui/material";
 
 export function Checkbox(props) {
-  const { label } = props;
+  const { label, onHandleValueChange, value, name } = props;
+
+  const convertToDefEventPara = (checkboxName, checkboxValue) => ({
+    target: {
+      name: checkboxName,
+      value: checkboxValue
+    }
+  });
+
   return (
     <FormGroup>
-      <FormControlLabel control={<MuiCheckbox />} label={label} />
+      <FormControlLabel
+        control={
+          <MuiCheckbox
+            name={name}
+            checked={value}
+            onChange={(e) =>
+              onHandleValueChange(convertToDefEventPara(name, e.target.checked))
+            }
+          />
+        }
+        label={label}
+      />
     </FormGroup>
   );
 }
